+++ CHANGES.md
@@ -21,6 +21,7 @@ Change Log
 * Added support for incrementally loading textures after a Model is ready. This allows the Model to be visible as soon as possible while its textures are loaded in the background.
 * Added `Cartographic.fromCartesian` function.
 * Added `Camera.getPixelSize` function to get the size of a pixel in meters based on the current view.
+* Added `ImageMaterialProperty.alpha` and a `alpha` uniform to `Image` `Material` types to control overall image opacity. It defaults to 1.0, fully opaque.
 
 ### 1.15 - 2015-11-02